Use minimum difficulty before DGW3 switch
- Do not bother calculating difficulty before switch to DGW3, because it
is impossible to validate difficulty of the original DGW algorithm. Nice
side effect: On regtest and simnet networks we can control number of
instamineble blocks. Difficulty rises very quickly if we use DGW3 from
the beginning. This happens due to small retarget window (of 24 blocks)
unlike Bitcoin where it is 2016 blocks.

- Set DGW3 switch block to 5000 on simnet and regtest. Set lower minimum
difficulty for regtest.

- Fix RPC tests.